Moved soundFurniture to its own package in model
