+++ src/main/java/app/model/soundFurniture/SoundFurnitureFactory.java
@@ -1,0 +1,13 @@
+package app.model.soundFurniture;
+
+import app.controller.settings.SettingsObject;
+import app.model.soundBoundary.SoundBoundaryFactory;
+
+public abstract class SoundFurnitureFactory {
+    public static SoundFurniture make(SettingsObject object)
+    {
+        SoundFurniture soundFurniture = new SoundFurnitureBase(SoundBoundaryFactory.make(object));
+
+        return soundFurniture;
+    }
+}
