+++ src/test/java/testing/SoundFurnitureTest.java
@@ -1,8 +1,8 @@
 package testing;
 
 import app.controller.linAlg.Vector;
-import app.controller.soundEngine.SoundBlockingBlock;
-import app.controller.soundEngine.SoundFurniture;
+import app.model.soundFurniture.SoundFurniture;
+import app.model.soundFurniture.SoundFurnitureBase;
 import javafx.geometry.Rectangle2D;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public class SoundFurnitureTest {
     @Test public void isInsideTest(){
-        SoundFurniture soundFurniture = new SoundBlockingBlock(new Rectangle2D(0,0,10,10));
+        SoundFurniture soundFurniture = new SoundFurnitureBase(new Rectangle2D(0,0,10,10));
 
         Vector p1 = new Vector(1,1);
         Vector p2 = new Vector(2,5);
@@ -22,7 +22,7 @@ public class SoundFurnitureTest {
     }
 
     @Test public void notIsInsideTest(){
-        SoundFurniture soundFurniture = new SoundBlockingBlock(new Rectangle2D(0,0,10,10));
+        SoundFurniture soundFurniture = new SoundFurnitureBase(new Rectangle2D(0,0,10,10));
 
         Vector p1 = new Vector(11,1);
         Vector p2 = new Vector(-1,1);
@@ -34,7 +34,7 @@ public class SoundFurnitureTest {
     }
 
     @Test public void onOutlineTest(){
-        SoundFurniture soundFurniture = new SoundBlockingBlock(new Rectangle2D(0,0,10,10));
+        SoundFurniture soundFurniture = new SoundFurnitureBase(new Rectangle2D(0,0,10,10));
 
         Vector p1 = new Vector(0,1);
         Vector p2 = new Vector(1,0);
@@ -51,7 +51,7 @@ public class SoundFurnitureTest {
     }
 
     @Test public void isInsideOutlineTest(){
-        SoundFurniture soundFurniture = new SoundBlockingBlock(new Rectangle2D(0,0,10,10));
+        SoundFurniture soundFurniture = new SoundFurnitureBase(new Rectangle2D(0,0,10,10));
 
         Vector p1 = new Vector(0,1);
         Vector p2 = new Vector(1,0);
